backend/AIProcessingService/workers/requirements.txt
# Core dependencies
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.11.0
httpx>=0.25.0
aio-pika>=9.3.0
boto3>=1.34.0

# OpenCV for image processing and mask refinement
opencv-python>=4.8.0

# ML Frameworks
torch>=2.1.0
torchvision>=0.16.0

# OpenAI CLIP
git+https://github.com/openai/CLIP.git

# Meta SAM2
git+https://github.com/facebookresearch/segment-anything-2.git

# Ultralytics YOLO
ultralytics>=8.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
